Correctly indent multiply nested BlockStatement
js2coffee was not correctly indenting code such as:

    while (true) {
          {
                {
                        var a = 0;
                        var b = 1;
                }
          }
    }

In that case you'd get:

    loop
        a = 0
      b = 1

The reason is that the code that emits 'BlockStatement' was
unconditionally prepending the current indent to each of its children.
In the preceding case, the code would add an indent to the top-level
BlockStatement then an indent for each child "a = 0" and "b = 0", resulting
in then output:

    ["loop", "\n", "  ", "  ", "a", "=", "0", "\n", " ", "b", "=", "0", "\n"]

This code makes it so that the code that emits 'BlockStatement' only outputs
an indent if the child isn't of type 'BlockStatement', effectively flattening
the output.