Properly reset scope member after popStack
The `scope` member in the transformer base was being set to
the previously pushed scope on `popStack`, instead of the
scope that was active that before the previously pushed scope
was pushed. This resulted in code like:

    var a = 0;
    var foo = function () {
      var b = function () {
        return 0;
      };
      var a = 1;
    };

Being compiled into:

    a = 0
    foo = ->
      b = ->
        `var a`
        0
      a = 1
      return

The preceding example would happen because the duplicate
VariableDeclaration for `a` would trigger js2coffee's variable shadowing
handling code, automatically placing the `var a` at the the
top of the current scope's body. The current scope was erroneously
left as the last pushed scope, hence `var a` would be placed at the
top of the "b" function.